requirements.txt
flask
requests
jsonpickle
pymysql
pyyaml
pypika